requirements.txt
admin-extra-urls==2.2.0
aiohttp==3.7.4
amqp==5.0.2
appnope==0.1.2
asgiref==3.3.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
Automat==0.8.0
autopep8==1.4.4
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==3.3.0
blinker==1.4
boto3==1.16.51
botocore==1.19.51
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cliapp==1.20180812.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloud-init==20.4
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect==1.1.0
curlify==2.2.1
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.12
diff-match-patch==20200713
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.4
django-admin-autocomplete-filter==0.6.1
django-admin-filters==0.1.3
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.6.3
django-admirarchy==1.2.1
django-allauth==0.39.1
django-appconf==1.0.4
django-autoslug==1.9.8
django-cacheops==5.1
django-celery-beat==2.1.0
django-celery-results==2.0.1
django-cities-light==3.8.1
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cms==3.8.0
django-common-helpers==0.9.2
django-cors-headers==3.6.0
django-countries==7.0
django-countries-plus==1.3.1
django-country-dialcode==0.6.3
django-country-dialcode-sb==0.6.4
django-crispy-forms==1.7.2
django-debug-toolbar==3.2
django-easy-select2==1.5.7
django-elasticsearch-dsl==7.1.4
django-extensions==3.1.0
django-extra-views==0.13.0
django-filer==2.0.2
django-filter==2.4.0
django-formtools==2.2
django-imagekit==4.0.2
django-import-export==2.4.0
django-js-asset==1.2.2
django-languages-plus==1.1.1
django-mathfilters==1.0.0
django-model-utils==4.1.1
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-polymorphic==3.0.0
django-quickbooks==0.6.4.2
django-redis==4.12.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-sekizai==2.0.0
django-solo==1.1.5
django-static-sitemaps==4.6.0
django-storages==1.11.1
django-timezone-field==4.1.1
django-tinymce==3.2.0
django-treebeard==4.4
django-url-filter==0.3.15
django-user-agents==0.4.0
django-utils==0.0.2
django-widget-tweaks==1.4.8
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-text-ckeditor==4.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
drf-yasg==1.20.0
easy-thumbnails==2.7.1
ecdsa==0.14.1
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
entrypoints==0.3
enum-compat==0.0.3
et-xmlfile==1.0.1
facebook-business==10.0.0
facebook-sdk==3.1.0
favicon==0.7.0
fire==0.4.0
funcy==1.15
future==0.18.2
geographiclib==1.50
geopy==2.1.0
google-services-api==0.0.8
gunicorn==20.0.4
haversine==2.3.1
html5lib==1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.10
importlib-metadata==3.3.0
incremental==16.10.1
indian-cities==0.0.4
inflection==0.5.1
intuit-oauth==1.2.2
ipykernel==4.6.1
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keyring==18.0.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
localflavor==1.8
lxml==4.4.2
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
memoize==1.0.0
meta==1.0.2
mistune==0.8.4
more-itertools==4.2.0
MouseInfo==0.1.3
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
num2words==0.5.10
numpy==1.19.5
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
optional-django==0.3.0
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parse==1.19.0
parso==0.8.1
pep8==1.7.1
pexpect==4.8.0
phonenumbers==8.12.15
pickleshare==0.7.5
pilkit==2.0
Pillow==8.0.1
portpicker==1.2.0
progressbar2==3.53.1
prompt-toolkit==3.0.10
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAutoGUI==0.9.52
pycodestyle==2.5.0
pycountry==20.7.3
pycparser==2.18
pydal==20200910.1
PyGetWindow==0.0.9
Pygments==2.7.3
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pyquery==1.4.3
PyRect==0.1.4
pyrsistent==0.15.5
PyScreeze==0.1.27
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-crontab==2.5.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.1
python-dotenv==0.15.0
python-jose==3.2.0
python-quickbooks==0.8.1
python-utils==2.4.0
python3-openid==3.1.0
python3-xlib==0.15
PyTweening==1.0.3
pytz==2020.5
pywebcopy==6.3.0
PyYAML==5.3.1
pyzmq==22.1.0
quickbooks-python==0.1.5
rauth==0.7.3
raven==6.10.0
razorpay==1.2.0
react==4.3.0
redis==3.5.3
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
SecretStorage==2.3.1
selenium==3.141.0
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.17.2
six==1.11.0
sos==4.0
soupsieve==2.1
speedtest-cli==2.1.2
spyne==2.13.12b0
sqlparse==0.4.1
ssh-import-id==5.10
stripe==2.27.0
systemd-python==234
tablib==3.0.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
tornado==4.5.3
traitlets==4.3.3
ttystatus==0.38
Twisted==18.9.0
typing-extensions==3.7.4.3
ua-parser==0.10.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.2
uritemplate==3.0.1
urllib3==1.26.2
user-agents==2.2.0
vine==5.0.0
vonage==2.5.5
w3lib==1.22.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
wget==3.2
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.3
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
zipp==3.4.0
zope.interface==4.7.1
